null check

+++ src/main/Database/File/FileDaoImpl.java
@@ -53,8 +53,12 @@ public class FileDaoImpl implements FileDao {
 
   @Override
   public void clear() {
-    fileCollection.drop();
-    fileBucket.drop();
+    if(fileCollection != null) {
+      fileCollection.drop();
+    }
+    if (fileBucket != null) {
+      fileBucket.drop();
+    }
   }
 
   @Override
